[#39] 카테고리 다건 수정 시 saveAll을 사용하도록 변경, 카테고리 조회 메서드 BlogService로 이동
