+++ src/main/java/com/plogcareers/backend/blog/domain/dto/SetCategoriesRequest.java
@@ -1,7 +1,7 @@
 package com.plogcareers.backend.blog.domain.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.plogcareers.backend.blog.domain.model.CategoryUpdateDTO;
+import com.plogcareers.backend.blog.domain.model.CreateCategoryDTO;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -13,7 +13,7 @@ import java.util.List;
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-public class UpdateCategoriesRequest {
+public class SetCategoriesRequest {
     @JsonProperty("categories")
-    private List<CategoryUpdateDTO> categoriesUpdateDTO;
+    private List<CreateCategoryDTO> categories;
 }
