+++ src/main/java/com/plogcareers/backend/blog/service/BlogService.java
@@ -1,10 +1,11 @@
 package com.plogcareers.backend.blog.service;
 
 import com.plogcareers.backend.blog.domain.dto.CreateCategoryRequest;
-import com.plogcareers.backend.blog.domain.dto.UpdateCategoriesRequest;
+import com.plogcareers.backend.blog.domain.dto.ListCategoryResponse;
+import com.plogcareers.backend.blog.domain.dto.SetCategoriesRequest;
 import com.plogcareers.backend.blog.domain.entity.Blog;
 import com.plogcareers.backend.blog.domain.entity.Category;
-import com.plogcareers.backend.blog.domain.model.CategoryUpdateDTO;
+import com.plogcareers.backend.blog.domain.model.CategoryDTO;
 import com.plogcareers.backend.blog.exception.BlogNotFoundException;
 import com.plogcareers.backend.blog.exception.CategoryDuplicatedException;
 import com.plogcareers.backend.blog.repository.BlogRepository;
@@ -14,6 +15,7 @@ import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
 import javax.validation.constraints.NotNull;
+import java.util.List;
 
 @Service
 @RequiredArgsConstructor
@@ -29,19 +31,29 @@ public class BlogService {
         categoryRepository.save(createCategoryRequest.toEntity(blog));
     }
 
+    // 카테고리 가져오기
+    public ListCategoryResponse listCategory(Long blogId) throws BlogNotFoundException {
+        if (blogRepository.existsById(blogId)) {
+            List<CategoryDTO> categoryList = categoryRepository.findCategoryByBlogIdOrderBySort(blogId)
+                    .stream()
+                    .map(Category::toCategoryDto)
+                    .toList();
+            return ListCategoryResponse.builder()
+                    .categories(categoryList)
+                    .build();
+        }
+        throw new BlogNotFoundException();
+    }
+
+    // 카테고리 다건 수정하기
     @Transactional
-    public void setCategory(Long blogId, @NotNull UpdateCategoriesRequest request) throws BlogNotFoundException {
+    public void setCategories(Long blogId, @NotNull SetCategoriesRequest request) throws BlogNotFoundException {
         Blog blog = blogRepository.findById(blogId).orElseThrow(BlogNotFoundException::new);
         categoryRepository.deleteCategoryByBlogId(blogId);
-        for (CategoryUpdateDTO categoryUpdateDTO : request.getCategoriesUpdateDTO()) {
-            categoryRepository.save(
-                    Category.builder()
-                            .categoryName(categoryUpdateDTO.getCategoryName())
-                            .categoryDesc(categoryUpdateDTO.getCategoryDesc())
-                            .sort(categoryUpdateDTO.getSort())
-                            .blog(blog)
-                            .build()
-            );
+        List<Category> categories = request.getCategories().stream()
+                        .map(createCategoryDTO -> createCategoryDTO.toEntity(blog))
+                        .toList();
+        categoryRepository.saveAll(categories);
         }
     }
-}
+
